+++ DEMToolbox/classes/particle_samples.py
@@ -8,7 +8,7 @@ class ParticleSamples():
     ----------
     name: (str)
         Name of the samples column
-    sample_attribute: (ParticleAttribute)
+    ParticleAttribute: (ParticleAttribute)
         ParticleAttribute object containing the particle ids and their
         corresponding sample ids
     cells: (np.ndarray)
@@ -17,12 +17,26 @@ class ParticleSamples():
         Array of cell ids that contain particles
     particles: (np.ndarray)
         Array of number of particles in each cell
+    n_cells: (int)
+        Number of possible cells
+    n_occupied_cells: (int)
+        Number of cells that contain particles
     n_sampled_particles: (int)
         Number of particles in the sampled cells
     n_unsampled_particles: (int)
         Number of particles not in the sampled cells
-    data: (np.ndarray)
-        Array of particle ids and their corresponding sample ids
+    vector_1_centers: (np.ndarray)
+        1D array of the vector 1 cell centers in the sample space
+    vector_1_bounds: (np.ndarray)
+        Array of the vector 1 bounds in the sample space
+    vector_2_centers: (np.ndarray)
+        1D array of the vector 2 cell centers in the sample space
+    vector_2_bounds: (np.ndarray)
+        Array of the vector 2 bounds in the sample space
+    vector_3_centers: (np.ndarray)
+        1D array of the vector 3 cell centers in the sample space
+    vector_3_bounds: (np.ndarray)
+        Array of the vector 3 bounds in the sample space
     """
     def __init__(self, 
                  name, 
@@ -31,7 +45,14 @@ class ParticleSamples():
                  occupied_cells,
                  particles, 
                  n_sampled_particles, 
-                 n_unsampled_particles):
+                 n_unsampled_particles,
+                 vector_1_centers=None,
+                 vector_1_bounds=None,
+                 vector_2_centers=None,
+                 vector_2_bounds=None,
+                 vector_3_centers=None,
+                 vector_3_bounds=None,
+                 ):
         
         self.name = name
         self.ParticleAttribute = sample_attribute
@@ -42,3 +63,9 @@ class ParticleSamples():
         self.n_occupied_cells = np.size(occupied_cells)
         self.n_sampled_particles = n_sampled_particles
         self.n_unsampled_particles = n_unsampled_particles
+        self.vector_1_centers = vector_1_centers
+        self.vector_1_bounds = vector_1_bounds
+        self.vector_2_centers = vector_2_centers
+        self.vector_2_bounds = vector_2_bounds
+        self.vector_3_centers = vector_3_centers
+        self.vector_3_bounds = vector_3_bounds